+++ drizzle/0074_flashy_king_cobra.sql
@@ -1,0 +1,1 @@
+ALTER TABLE "automations" ADD COLUMN "disabled_reason" text;